+++ forge20/src/main/java/com/vecoo/extraquests/ExtraQuests.java
@@ -1,5 +1,6 @@
 package com.vecoo.extraquests;
 
+import com.mojang.logging.LogUtils;
 import com.vecoo.extraquests.command.ExtraQuestsCommand;
 import com.vecoo.extraquests.config.LocaleConfig;
 import com.vecoo.extraquests.config.ServerConfig;
@@ -18,17 +19,15 @@ import net.minecraftforge.event.RegisterCommandsEvent;
 import net.minecraftforge.event.server.ServerStartedEvent;
 import net.minecraftforge.event.server.ServerStartingEvent;
 import net.minecraftforge.event.server.ServerStoppingEvent;
-import net.minecraftforge.eventbus.api.EventPriority;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.server.permission.events.PermissionGatherEvent;
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
+import org.slf4j.Logger;
 
 @Mod(ExtraQuests.MOD_ID)
 public class ExtraQuests {
     public static final String MOD_ID = "extraquests";
-    private static final Logger LOGGER = LogManager.getLogger("ExtraQuests");
+    private static final Logger LOGGER = LogUtils.getLogger();
 
     private static ExtraQuests instance;
 
@@ -50,9 +49,7 @@ public class ExtraQuests {
 
     @SubscribeEvent
     public void onPermissionGather(PermissionGatherEvent.Nodes event) {
-        if (!event.getNodes().contains(PermissionNodes.EXTRAQUESTS_COMMAND)) {
-            event.addNodes(PermissionNodes.EXTRAQUESTS_COMMAND);
-        }
+        PermissionNodes.registerPermission(event);
     }
 
     @SubscribeEvent
@@ -70,7 +67,7 @@ public class ExtraQuests {
         loadStorage();
     }
 
-    @SubscribeEvent(priority = EventPriority.LOWEST)
+    @SubscribeEvent
     public void onServerStopping(ServerStoppingEvent event) {
         this.timerProvider.write();
     }
@@ -82,23 +79,23 @@ public class ExtraQuests {
             this.locale = new LocaleConfig();
             this.locale.init();
         } catch (Exception e) {
-            LOGGER.error("[ExtraQuests] Error load config.", e);
+            LOGGER.error("Error load config.", e);
         }
     }
 
     public void loadStorage() {
         try {
-            if (timerProvider == null) {
+            if (this.timerProvider == null) {
                 this.timerProvider = new TimerProvider("/%directory%/storage/ExtraQuests/", this.server);
             }
 
             this.timerProvider.init();
         } catch (Exception e) {
-            LOGGER.error("[ExtraQuests] Error load storage.", e);
+            LOGGER.error("Error load storage.", e);
         }
     }
 
-    public void registerQuests() {
+    private void registerQuests() {
         KeyValueTask.TYPE = TaskTypes.register(new ResourceLocation(ExtraQuests.MOD_ID, "key_value"), KeyValueTask::new, () -> Icon.getIcon("minecraft:item/paper"));
         KeyValueReward.TYPE = RewardTypes.register(new ResourceLocation(ExtraQuests.MOD_ID, "key_value"), KeyValueReward::new, () -> Icon.getIcon("minecraft:item/paper"));
         TimerReward.TYPE = RewardTypes.register(new ResourceLocation(ExtraQuests.MOD_ID, "timer"), TimerReward::new, () -> Icon.getIcon("minecraft:item/clock_07"));