+++ forge20/src/main/java/com/vecoo/extraquests/storage/TimerStorage.java
@@ -1,5 +1,7 @@
 package com.vecoo.extraquests.storage;
 
+import org.jetbrains.annotations.NotNull;
+
 import java.util.Objects;
 import java.util.UUID;
 
@@ -8,16 +10,18 @@ public class TimerStorage {
     private final String questID;
     private final long endTime;
 
-    public TimerStorage(UUID playerUUID, String questID, int secondsTimer) {
+    public TimerStorage(@NotNull UUID playerUUID, @NotNull String questID, int secondsTimer) {
         this.playerUUID = playerUUID;
         this.questID = questID;
         this.endTime = System.currentTimeMillis() + (secondsTimer * 1000L);
     }
 
+    @NotNull
     public UUID getPlayerUUID() {
         return this.playerUUID;
     }
 
+    @NotNull
     public String getQuestID() {
         return this.questID;
     }