+++ neoforge21/src/main/java/com/vecoo/extraquests/ExtraQuests.java
@@ -1,5 +1,6 @@
 package com.vecoo.extraquests;
 
+import com.mojang.logging.LogUtils;
 import com.vecoo.extraquests.command.ExtraQuestsCommand;
 import com.vecoo.extraquests.config.LocaleConfig;
 import com.vecoo.extraquests.config.ServerConfig;
@@ -13,7 +14,6 @@ import dev.ftb.mods.ftbquests.quest.reward.RewardTypes;
 import dev.ftb.mods.ftbquests.quest.task.TaskTypes;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.MinecraftServer;
-import net.neoforged.bus.api.EventPriority;
 import net.neoforged.bus.api.SubscribeEvent;
 import net.neoforged.fml.common.Mod;
 import net.neoforged.neoforge.common.NeoForge;
@@ -22,13 +22,13 @@ import net.neoforged.neoforge.event.server.ServerStartedEvent;
 import net.neoforged.neoforge.event.server.ServerStartingEvent;
 import net.neoforged.neoforge.event.server.ServerStoppingEvent;
 import net.neoforged.neoforge.server.permission.events.PermissionGatherEvent;
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
+import org.slf4j.Logger;
+
 
 @Mod(ExtraQuests.MOD_ID)
 public class ExtraQuests {
     public static final String MOD_ID = "extraquests";
-    private static final Logger LOGGER = LogManager.getLogger("ExtraQuests");
+    private static final Logger LOGGER = LogUtils.getLogger();
 
     private static ExtraQuests instance;
 
@@ -50,9 +50,7 @@ public class ExtraQuests {
 
     @SubscribeEvent
     public void onPermissionGather(PermissionGatherEvent.Nodes event) {
-        if (!event.getNodes().contains(PermissionNodes.EXTRAQUESTS_COMMAND)) {
-            event.addNodes(PermissionNodes.EXTRAQUESTS_COMMAND);
-        }
+        PermissionNodes.registerPermission(event);
     }
 
     @SubscribeEvent
@@ -70,7 +68,7 @@ public class ExtraQuests {
         loadStorage();
     }
 
-    @SubscribeEvent(priority = EventPriority.LOWEST)
+    @SubscribeEvent
     public void onServerStopping(ServerStoppingEvent event) {
         this.timerProvider.write();
     }
@@ -82,23 +80,23 @@ public class ExtraQuests {
             this.locale = new LocaleConfig();
             this.locale.init();
         } catch (Exception e) {
-            LOGGER.error("[ExtraQuests] Error load config.", e);
+            LOGGER.error("Error load config.", e);
         }
     }
 
     public void loadStorage() {
         try {
-            if (timerProvider == null) {
+            if (this.timerProvider == null) {
                 this.timerProvider = new TimerProvider("/%directory%/storage/ExtraQuests/", this.server);
             }
 
             this.timerProvider.init();
         } catch (Exception e) {
-            LOGGER.error("[ExtraQuests] Error load storage.", e);
+            LOGGER.error("Error load storage.", e);
         }
     }
 
-    public void registerQuests() {
+    private void registerQuests() {
         KeyValueTask.TYPE = TaskTypes.register(ResourceLocation.fromNamespaceAndPath(ExtraQuests.MOD_ID, "key_value"), KeyValueTask::new, () -> Icon.getIcon("minecraft:item/paper"));
         KeyValueReward.TYPE = RewardTypes.register(ResourceLocation.fromNamespaceAndPath(ExtraQuests.MOD_ID, "key_value"), KeyValueReward::new, () -> Icon.getIcon("minecraft:item/paper"));
         TimerReward.TYPE = RewardTypes.register(ResourceLocation.fromNamespaceAndPath(ExtraQuests.MOD_ID, "timer"), TimerReward::new, () -> Icon.getIcon("minecraft:item/clock_07"));